Add velocity read from UBX

+++ bsp_cc26x2/clocks/clocks.h
@@ -22,7 +22,7 @@ extern "C" {
 #define MIN_2_MS(MIN) ((MIN)*60U * 1000U)
 #define S_2_MS(S) ((S)*1000U)
 #define S_2_US(S) ((S)*1000000U)
-#define MS_2_S(ms) (ms / 1000)
+#define MS_2_S(ms) ((((double)ms) / 1000.0))
 #define MS_2_MIN(ms) (ms / (1000 * 60))
 
 uint32_t get_time_ms32(void);
